LayoutTests/inspector/editor/text-editor-model-replace-tabs.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>

function test()
{
    var textModel = new WebInspector.TextEditorModel();
    textModel.replaceTabsWithSpaces = true;
    var testCases = new Array("","\t","A\t","AB\t","ABC\t","\tA","A\tB","\t\tA\t","\tA\t\t","A","\tABCD");
    for (var i = 0; i < testCases.length; i++) {
        var input = testCases[i];
        var output = new Array(input);
        textModel._replaceTabsIfNeeded(output);
        output = output[0];
        InspectorTest.addResult("Test Case #" + i);
        InspectorTest.addResult("Input:  " + input.replace(/\t/g, "\\t").replace(/ /g, "."));
        InspectorTest.addResult("Output: " + output.replace(/\t/g, "\\t").replace(/ /g, "."));
        InspectorTest.addResult("");
    }

    var input = [];
    var output = [];
    for (var i = 0; i < 10000; i++) {
        input.push("\t\t\t\t\t\t\t\t\t\t"); // 10 tabulations
        output.push("                                        "); // 40 spaces
    }
    input = new Array(input.join(""));
    output = output.join("");
    textModel._replaceTabsIfNeeded(input);
    var correctResult = (input[0] === output);
    InspectorTest.addResult("Very Long test: " + (correctResult ? "PASS" : "FAIL"));

    InspectorTest.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>
This test checks the text editor model _replaceTabsIfNeeded method.
In test cases tabulations are denoted with "\t" and spaces are rendered as ".";
</p>
</body>
</html>
