LayoutTests/inspector/editor/text-editor-model-replace-tabs-expected.txt

This test checks the text editor model _replaceTabsIfNeeded method. In test cases tabulations are denoted with "\t" and spaces are rendered as ".";

Test Case #0
Input:  
Output: 

Test Case #1
Input:  \t
Output: ....

Test Case #2
Input:  A\t
Output: A...

Test Case #3
Input:  AB\t
Output: AB..

Test Case #4
Input:  ABC\t
Output: ABC.

Test Case #5
Input:  \tA
Output: ....A

Test Case #6
Input:  A\tB
Output: A...B

Test Case #7
Input:  \t\tA\t
Output: ........A...

Test Case #8
Input:  \tA\t\t
Output: ....A.......

Test Case #9
Input:  A
Output: A

Test Case #10
Input:  \tABCD
Output: ....ABCD

Very Long test: PASS
